requirements.txt
streamlit
langchain-classic
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
python-dotenv
pydantic